conda_env/segmentation_methods/optimal_flow-0.1.11/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alabaster=0.7.13=pypi_0
altgraph=0.17.4=pypi_0
arrow=1.3.0=pypi_0
astroid=3.0.1=pypi_0
asttokens=2.4.1=pypi_0
atomicwrites=1.4.1=pypi_0
attrs=23.1.0=pypi_0
auto-py-to-exe=2.42.0=pypi_0
autopep8=2.0.4=pypi_0
babel=2.13.1=pypi_0
bcrypt=4.1.1=pypi_0
beautifulsoup4=4.12.2=pypi_0
binaryornot=0.4.4=pypi_0
black=23.11.0=pypi_0
bleach=6.1.0=pypi_0
blinker=1.7.0=pypi_0
bottle=0.12.25=pypi_0
bottle-websocket=0.2.9=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.08.22=haa95532_0
category-encoders=2.6.3=pypi_0
certifi=2023.11.17=pypi_0
cffi=1.16.0=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.0=pypi_0
contourpy=1.2.0=pypi_0
cookiecutter=2.5.0=pypi_0
cryptography=41.0.7=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
diff-match-patch=20230430=pypi_0
dill=0.3.7=pypi_0
docstring-to-markdown=0.13=pypi_0
docutils=0.20.1=pypi_0
eel=0.16.0=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.19.0=pypi_0
flake8=6.1.0=pypi_0
flask=3.0.0=pypi_0
fonttools=4.45.1=pypi_0
future=0.18.3=pypi_0
gevent=23.9.1=pypi_0
gevent-websocket=0.10.1=pypi_0
greenlet=3.0.1=pypi_0
idna=3.6=pypi_0
imagesize=1.4.1=pypi_0
importlib-metadata=6.8.0=pypi_0
inflection=0.5.1=pypi_0
intervaltree=3.1.0=pypi_0
ipykernel=6.27.1=pypi_0
ipython=8.18.1=pypi_0
isort=5.12.0=pypi_0
itsdangerous=2.1.2=pypi_0
jaraco-classes=3.3.0=pypi_0
jedi=0.19.1=pypi_0
jellyfish=1.0.3=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.11.1=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-core=5.5.0=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
keyring=24.3.0=pypi_0
kiwisolver=1.4.5=pypi_0
libffi=3.4.4=hd77b12b_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mccabe=0.7.0=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
more-itertools=10.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclient=0.9.0=pypi_0
nbconvert=7.11.0=pypi_0
nbformat=5.9.2=pypi_0
nest-asyncio=1.5.8=pypi_0
numpy=1.26.2=pypi_0
numpydoc=1.6.0=pypi_0
openssl=3.0.12=h2bbff1b_0
optimalflow=0.1.11=pypi_0
packaging=23.2=pypi_0
pandas=1.5.3=pypi_0
pandocfilters=1.5.0=pypi_0
paramiko=3.3.1=pypi_0
parso=0.8.3=pypi_0
pathspec=0.11.2=pypi_0
patsy=0.5.3=pypi_0
pefile=2023.2.7=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.1.0=pypi_0
pip=23.3.1=py310haa95532_0
platformdirs=4.0.0=pypi_0
plotly=5.18.0=pypi_0
pluggy=1.3.0=pypi_0
prompt-toolkit=3.0.41=pypi_0
psutil=5.9.6=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pycodestyle=2.11.1=pypi_0
pycparser=2.21=pypi_0
pydocstyle=6.3.0=pypi_0
pyflakes=3.1.0=pypi_0
pygments=2.17.2=pypi_0
pyinstaller=6.2.0=pypi_0
pyinstaller-hooks-contrib=2023.10=pypi_0
pylint=3.0.2=pypi_0
pylint-venv=3.0.3=pypi_0
pyls-spyder=0.4.0=pypi_0
pynacl=1.5.0=pypi_0
pyparsing=3.1.1=pypi_0
pyqt5=5.15.10=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.13.0=pypi_0
pyqtwebengine=5.15.6=pypi_0
pyqtwebengine-qt5=5.15.2=pypi_0
python=3.10.13=he1021f5_0
python-dateutil=2.8.2=pypi_0
python-lsp-black=1.3.0=pypi_0
python-lsp-jsonrpc=1.1.2=pypi_0
python-lsp-server=1.9.0=pypi_0
python-slugify=8.0.1=pypi_0
pytoolconfig=1.2.6=pypi_0
pytz=2023.3.post1=pypi_0
pywin32=306=pypi_0
pywin32-ctypes=0.2.2=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.1=pypi_0
qdarkstyle=3.2.3=pypi_0
qstylizer=0.2.2=pypi_0
qtawesome=1.2.3=pypi_0
qtconsole=5.5.1=pypi_0
qtpy=2.4.1=pypi_0
referencing=0.31.0=pypi_0
requests=2.31.0=pypi_0
rich=13.7.0=pypi_0
rope=1.11.0=pypi_0
rpds-py=0.13.1=pypi_0
rtree=1.1.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.4=pypi_0
setuptools=68.0.0=py310haa95532_0
six=1.16.0=pypi_0
snowballstemmer=2.2.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.5=pypi_0
sphinx=7.2.6=pypi_0
sphinxcontrib-applehelp=1.0.7=pypi_0
sphinxcontrib-devhelp=1.0.5=pypi_0
sphinxcontrib-htmlhelp=2.0.4=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=1.0.6=pypi_0
sphinxcontrib-serializinghtml=1.1.9=pypi_0
spyder=5.5.0=pypi_0
spyder-kernels=2.5.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack-data=0.6.3=pypi_0
statsmodels=0.14.0=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.2.3=pypi_0
text-unidecode=1.3=pypi_0
textdistance=4.6.0=pypi_0
threadpoolctl=3.2.0=pypi_0
three-merge=0.1.1=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h2bbff1b_0
tomli=2.0.1=pypi_0
tomlkit=0.12.3=pypi_0
tornado=6.4=pypi_0
traitlets=5.14.0=pypi_0
types-python-dateutil=2.8.19.14=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2023c=h04d1e81_0
ujson=5.8.0=pypi_0
urllib3=2.1.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=3.0.0=pypi_0
wcwidth=0.2.12=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=3.0.1=pypi_0
whatthepatch=1.0.5=pypi_0
wheel=0.41.2=py310haa95532_0
whichcraft=0.6.1=pypi_0
wtforms=3.1.1=pypi_0
xz=5.4.2=h8cc25b3_0
yapf=0.40.2=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h8cc25b3_0
zope-event=5.0=pypi_0
zope-interface=6.1=pypi_0
